+++ src/header/neuronNetwork.hpp
@@ -35,7 +35,7 @@ public:
      * @param inputs A vector of input values.
      * @return The output of the network. 
      */
-    std::vector<int> feedForward(const std::vector<int>& inputs);
+    std::vector<float> feedForward(const std::vector<float>& inputs);
     
     /**
      * @brief Prints the network details.
